third_party/blink/web_tests/fast/table/percent-widths-float.html

<!DOCTYPE HTML>
<style>
    table { border-collapse: collapse; width: 1000px;}
    #green { background-color: green; padding: 0; width: 1.5%;}
    #blue { background-color: blue; padding: 0; width: 0.5%;}
    #orange { background-color: orange; padding: 0; width: 10%; }
    div { height: 10px; }
</style>
<script src="../../resources/check-layout.js"></script>
<table>
    <td id="green" data-expected-width=125><div></div></td>
    <td id="blue" data-expected-width=42><div></div></td>
    <td id="orange" data-expected-width=833><div></div></td>
</table>
<p> crbug.com/476370: Test cells get appropriate width when we use fractional percentage values. </p>
<div id="console"></div>
<script>
    checkLayout('td', document.getElementById('console'));
</script>
