third_party/blink/web_tests/flag-specific/disable-layout-ng/fast/table/percent-widths-float-expected.txt

	
	

crbug.com/476370: Test cells get appropriate width when we use fractional percentage values.

FAIL:
Expected 833 for width, but got 834. 

<td id="orange" data-expected-width="833"><div></div></td>
FAIL:
Expected 42 for width, but got 41. 

<td id="blue" data-expected-width="42"><div></div></td>
PASS
